requirements.txt
appnope==0.1.2
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
bleach==3.3.1
boto3==1.18.23
botocore==1.21.23
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.7
Django==3.1.7
django-bleach==0.7.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-messages==0.6.0
django-mptt==0.12.0
django-mptt-admin==2.1.0
django-multiselectfield==0.1.12
django-polymorphic==3.0.0
django-polymorphic-tree==2.0
django-postman==4.2
django-resized==0.3.11
django-storages==1.11.1
django-tag-parser==3.2
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
idna==2.10
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jellyfish==0.8.2
jmespath==0.10.0
lib50==2.0.9
markdown2==2.4.0
matplotlib-inline==0.1.2
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
prompt-toolkit==3.0.18
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
submit50==3.0.3
termcolor==1.1.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
